fetch_text/requirements.txt
requests
google-cloud-storage
google-cloud-pubsub
flask
